A JDK1.5 compilation issue fixed.
git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk/carrot2@540 7ff1d41c-760d-0410-a7ff-a3a56f310b35

+++ applications/carrot2-odp-tools/src/com/stachoodev/carrot/odp/tools/TopicDumper.java
@@ -118,7 +118,7 @@ public class TopicDumper
                 }
                 else
                 {
-                    return ((Integer) l1[1]).compareTo(l2[1]);
+                    return ((Integer) l1[1]).compareTo((Integer)l2[1]);
                 }
             }
         });
